dnd_session_toolchain/requirements.txt
faster-whisper>=1.0.3
python-dotenv>=1.0.1
requests>=2.32.0
